backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.6
django-redis==5.4.0
djangorestframework==3.15.1
greenlet==3.0.3
idna==3.7
kombu==5.3.7
prompt_toolkit==3.0.45
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.4
requests==2.32.2
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.30
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
